+++ dev/com.ibm.ws.microprofile.reactive.messaging_fat/fat/src/com/ibm/ws/microprofile/reactive/messaging/fat/suite/KafkaMessagingTest.java
@@ -12,7 +12,6 @@ package com.ibm.ws.microprofile.reactive.messaging.fat.suite;
 
 import static com.ibm.ws.microprofile.reactive.messaging.fat.suite.ConnectorProperties.simpleIncomingChannel;
 import static com.ibm.ws.microprofile.reactive.messaging.fat.suite.ConnectorProperties.simpleOutgoingChannel;
-import static com.ibm.ws.microprofile.reactive.messaging.fat.suite.FATSuite.kafkaContainer;
 import static com.ibm.ws.microprofile.reactive.messaging.fat.suite.KafkaUtils.kafkaClientLibs;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.contains;
@@ -70,8 +69,8 @@ public class KafkaMessagingTest {
     @BeforeClass
     public static void setup() throws Exception {
         PropertiesAsset config = new PropertiesAsset()
-                        .include(simpleOutgoingChannel(kafkaContainer, "test-out"))
-                        .include(simpleIncomingChannel(kafkaContainer, "test-in", "test-consumer"));
+                        .include(simpleOutgoingChannel(PlaintextTests.kafkaContainer, "test-out"))
+                        .include(simpleIncomingChannel(PlaintextTests.kafkaContainer, "test-in", "test-consumer"));
 
         WebArchive war = ShrinkWrap.create(WebArchive.class, APP_NAME + ".war")
                         .addPackage(BasicMessagingBean.class.getPackage())
@@ -86,13 +85,13 @@ public class KafkaMessagingTest {
     @BeforeClass
     public static void initializeKafkaClients() {
         Map<String, Object> consumerConfig = new HashMap<>();
-        consumerConfig.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, kafkaContainer.getBootstrapServers());
+        consumerConfig.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, PlaintextTests.kafkaContainer.getBootstrapServers());
         consumerConfig.put(ConsumerConfig.GROUP_ID_CONFIG, "testClient");
         consumerConfig.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest");
         kafkaConsumer = new KafkaConsumer<>(consumerConfig, new StringDeserializer(), new StringDeserializer());
 
         Map<String, Object> producerConfig = new HashMap<>();
-        producerConfig.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, kafkaContainer.getBootstrapServers());
+        producerConfig.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, PlaintextTests.kafkaContainer.getBootstrapServers());
         kafkaProducer = new KafkaProducer<>(producerConfig, new StringSerializer(), new StringSerializer());
     }
